+++ check/controller/bootstrap.php
@@ -30,7 +30,7 @@ class Bootstrap
 
 		session_start();
 
-		define('CONTAO_CHECK_VERSION', '4.2');
+		define('CONTAO_CHECK_VERSION', '5.0');
 		define('IS_WINDOWS', (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN'));
 
 		$this->setLocale($this->getLanguage());
@@ -155,9 +155,9 @@ class Bootstrap
 			return;
 		}
 
-		putenv("LC_ALL=$locale");
+		putenv("LANG=$locale");
 		setlocale(LC_ALL, $locale);
-		bindtextdomain('messages', dirname(__FILE__) . '/locale');
+		bindtextdomain('messages', dirname(dirname(__FILE__)) . '/locale');
 		textdomain('messages');
 		bind_textdomain_codeset('messages', 'UTF-8');
 	}